Crash Fix for deinitBuffering() on Monochrome Calculators
If you are using n2DLib with a monochrome calculator, you may run in some crashes when you call deinitBufferring(). That's because n2DLib's monochrome handling is done before checking if the screen has been swapped in deinitBuffering() function. It should be after. So, that said, the fix is very simple. Just need change some lines in deinitBuffering().